api/requirements.txt
flask
flask-session
python-jose[cryptography]
passlib[bcrypt]
msal